requirements.txt
django
djangorestframework
django-cors-headers
psycopg2-binary
django-environ
Pillow
cloudinary
django-cloudinary-storage